added and tested email functionality

+++ app.rb
@@ -6,6 +6,7 @@ require 'pg'
 require 'cgi'
 require_relative 'lib/peep'
 require_relative 'lib/user'
+require_relative 'lib/send_email'
 
 ActiveRecord::Base.establish_connection(
   adapter:  'postgresql',
@@ -24,17 +25,15 @@ class Application < Sinatra::Base
   def self.escape_html_all_params(params) # santises all inputs
     params.each { |k, v| params[k] = CGI.escapeHTML(v) }
   end
-  # get /\/albums\/([0-9]+)/ do
 
-  get '/' do # homepage
-    p 1
+  get '/' do
     @user_id = session[:user_id]
     @username = @user_id.nil? ? nil : User.find(@user_id).username
     @peeps = Peep.last(20).sort_by { |peep| peep.date_time_created }.reverse
     erb(:index)
   end
 
-  get '/login' do # Login page
+  get '/login' do
     redirect to '/' unless session[:user_id].nil?
     @user_id = session[:user_id]
     @sign_up = params[:sign_up] == 'true'
@@ -48,13 +47,13 @@ class Application < Sinatra::Base
       params[:password].nil? || params[:username].nil?
     Application.escape_html_all_params(params)
     user = User.find_by(username: params[:username])
-    redirect to '/login&username_error=true' if user.nil?
-    redirect to '/login&password_error=true' if params[:password] != user.password
+    redirect to '/login?username_error=true' if user.nil?
+    redirect to '/login?password_error=true' if params[:password] != user.password
     session[:user_id] = user.id
     redirect to '/'
   end
 
-  post '/log-off' do
+  get '/log-off' do
     session.delete(:user_id)
     erb(:log_off)
   end
@@ -71,17 +70,17 @@ class Application < Sinatra::Base
     return status 400 if !session[:user_id].nil? || params[:name].nil? || params[:email].nil? ||
       params[:username].nil? || params[:password].nil?
     Application.escape_html_all_params(params)
-    error_params = ""
-    error_params += '&email_error=true' unless User.find_by(email: params[:email]).nil?
-    error_params += '&username_error=true' unless User.find_by(username: params[:username]).nil?
-    redirect to "/sign-up/new#{error_params}" unless error_params.empty?
+    error_params = []
+    error_params << 'email_error=true' unless User.find_by(email: params[:email]).nil?
+    error_params << 'username_error=true' unless User.find_by(username: params[:username]).nil?
+    redirect to "/sign-up/new#{'?' + error_params.join('&')}" unless error_params.empty?
     User.create(
       name: params[:name],
       username: params[:username],
       email: params[:email],
       password: params[:password]
     )
-    redirect to '/login&sign-up=true'
+    redirect to '/login?sign_up=true'
   end
 
   get '/peep/new' do
@@ -93,23 +92,40 @@ class Application < Sinatra::Base
     Application.escape_html_all_params(params)
     return 400 if session[:user_id].nil? || params[:content].nil?
     session[:time_admin] = Time.now if session[:time_admin].nil?
-    Peep.create(
+    peep = Peep.create(
       content: params[:content],
       date_time_created: session[:time_admin],
       user_id: session[:user_id]
     )
+    email_users(peep)
     redirect to '/'
   end
 
   post '/time-admin' do
     Application.escape_html_all_params(params)
     return 400 unless params[:password] == 'admin'
-    if params[:time] == 'reset'
-      session.delete(:time_admin)
-      return "Time reset"
-    end
+    return session.delete(:time_admin) if params[:time] == 'reset'
     time = params[:time].split(',').map(&:to_i)
     session[:time_admin] = Time.new(time[0], time[1], time[2], time[3], time[4], time[5])
     "Time set to #{session[:time_admin]}"
+  end
+
+  private
+  
+  def email_users(peep)
+    # sends users an email if they are mentioned in the peep
+    peep.content.split.each do |word|
+      to_contact = User.find_by(username: word[1, word.size])
+      if word[0] == '@' && !to_contact.nil?
+        SendEmail.new(nil, to_contact.email).send_email(email_message(peep))
+      end
+    end
+  end
+
+  def email_message(peep)
+    # creates email message
+    username = "You've been mentioned in a Peep by @#{User.find(peep.user_id).username}!"
+    content_date = "They wrote '#{peep.content}' on #{peep.date_time_created_formatted_string}."
+    username + " " + content_date
   end
 end
